# api/openapi.yaml
openapi: 3.0.3
info:
  title: Salão SaaS API
  description: API REST para gestão multi-tenant de salões de beleza e barbearias.
  version: 1.0.0
  contact:
    name: Salão SaaS

servers:
  - url: http://localhost:8080/api/v1
    description: Desenvolvimento local

tags:
  - name: Auth
    description: Autenticação e tokens JWT
  - name: Tenants
    description: Gestão de salões (tenants)
  - name: Services
    description: Serviços oferecidos
  - name: Appointments
    description: Agendamentos
  - name: Dashboard
    description: Métricas operacionais

paths:
  /health:
    get:
      tags: [Auth]
      summary: Health check da API
      security: []
      responses:
        '200':
          description: API operacional
          content:
            application/json:
              schema:
                type: object
                required: [status, service]
                properties:
                  status:
                    type: string
                    example: UP
                  service:
                    type: string
                    example: salao-api

  /auth/register:
    post:
      tags: [Auth]
      summary: Registra salão e usuário admin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Salão e admin criados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Dados invalidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email ou slug ja cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags: [Auth]
      summary: Autentica usuário e retorna JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciais invalidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/me:
    get:
      tags: [Auth]
      summary: Retorna usuário autenticado, tenant e role
      responses:
        '200':
          description: Perfil do usuário autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Token ausente ou invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Acesso negado

  /services:
    get:
      tags: [Services]
      summary: Lista serviços do tenant
      responses:
        '200':
          description: Lista de serviços
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceResponse'
    post:
      tags: [Services]
      summary: Cria serviço (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          description: Serviço criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
        '403':
          description: Acesso negado

  /services/{publicId}:
    get:
      tags: [Services]
      summary: Obtém serviço por publicId
      parameters:
        - $ref: '#/components/parameters/PublicId'
      responses:
        '200':
          description: Serviço encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
        '404':
          description: Serviço não encontrado
    put:
      tags: [Services]
      summary: Atualiza serviço (ADMIN)
      parameters:
        - $ref: '#/components/parameters/PublicId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceRequest'
      responses:
        '200':
          description: Serviço atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
        '403':
          description: Acesso negado
        '404':
          description: Serviço não encontrado
    delete:
      tags: [Services]
      summary: Desativa serviço (ADMIN, soft delete)
      parameters:
        - $ref: '#/components/parameters/PublicId'
      responses:
        '204':
          description: Serviço desativado
        '403':
          description: Acesso negado
        '404':
          description: Serviço não encontrado

  /appointments:
    get:
      tags: [Appointments]
      summary: Lista agendamentos (filtrado por role)
      responses:
        '200':
          description: Lista de agendamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentResponse'
    post:
      tags: [Appointments]
      summary: Cria agendamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
      responses:
        '201':
          description: Agendamento criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '400':
          description: Serviço inativo
        '409':
          description: Conflito de agenda

  /appointments/{publicId}/cancel:
    patch:
      tags: [Appointments]
      summary: Cancela agendamento
      parameters:
        - $ref: '#/components/parameters/PublicId'
      responses:
        '200':
          description: Agendamento cancelado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '403':
          description: Acesso negado
        '404':
          description: Agendamento não encontrado

  /dashboard/summary:
    get:
      tags: [Dashboard]
      summary: Resumo operacional do dia (UTC)
      responses:
        '200':
          description: Métricas do dia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummaryResponse'

components:
  parameters:
    PublicId:
      name: publicId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    RegisterRequest:
      type: object
      required: [salonName, slug, name, email, password]
      properties:
        salonName:
          type: string
          example: Studio Bella
        slug:
          type: string
          example: studio-bella
        name:
          type: string
          example: Maria Admin
        email:
          type: string
          format: email
          example: admin@studio.com
        password:
          type: string
          format: password
          minLength: 8
          example: senha1234

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@studio.com
        password:
          type: string
          format: password
          example: senha1234

    AuthResponse:
      type: object
      required: [token, user, tenant, role]
      properties:
        token:
          type: string
          description: JWT Bearer token
        user:
          $ref: '#/components/schemas/UserSummary'
        tenant:
          $ref: '#/components/schemas/TenantSummary'
        role:
          type: string
          enum: [ADMIN, PROFESSIONAL, CLIENT]

    UserSummary:
      type: object
      required: [publicId, email, name]
      properties:
        publicId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string

    TenantSummary:
      type: object
      required: [publicId, name, slug]
      properties:
        publicId:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string

    MeResponse:
      type: object
      required: [user, tenant, role]
      properties:
        user:
          $ref: '#/components/schemas/UserSummary'
        tenant:
          $ref: '#/components/schemas/TenantSummary'
        role:
          type: string
          enum: [ADMIN, PROFESSIONAL, CLIENT]

    ServiceResponse:
      type: object
      required: [publicId, name, durationMinutes, price, active]
      properties:
        publicId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        price:
          type: number
          format: decimal
        active:
          type: boolean

    CreateServiceRequest:
      type: object
      required: [name, durationMinutes, price]
      properties:
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          minimum: 1
        price:
          type: number
          format: decimal
          minimum: 0

    UpdateServiceRequest:
      type: object
      required: [name, durationMinutes, price, active]
      properties:
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          minimum: 1
        price:
          type: number
          format: decimal
          minimum: 0
        active:
          type: boolean

    CreateAppointmentRequest:
      type: object
      required: [servicePublicId, professionalPublicId, startAt]
      properties:
        servicePublicId:
          type: string
          format: uuid
        professionalPublicId:
          type: string
          format: uuid
        clientPublicId:
          type: string
          format: uuid
          description: Obrigatório para ADMIN/PROFESSIONAL; ignorado para CLIENT
        startAt:
          type: string
          format: date-time

    AppointmentResponse:
      type: object
      required: [publicId, service, professional, client, startAt, endAt, status, createdAt]
      properties:
        publicId:
          type: string
          format: uuid
        service:
          $ref: '#/components/schemas/UuidRef'
        professional:
          $ref: '#/components/schemas/UuidRef'
        client:
          $ref: '#/components/schemas/UuidRef'
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [SCHEDULED, CONFIRMED, CANCELLED, COMPLETED]
        createdAt:
          type: string
          format: date-time

    DashboardSummaryResponse:
      type: object
      required: [appointmentsToday, estimatedRevenue, occupancyRate]
      properties:
        appointmentsToday:
          type: integer
          format: int64
        estimatedRevenue:
          type: number
          format: decimal
        occupancyRate:
          type: number
          format: double
          minimum: 0
          maximum: 1

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Campo obrigatório ausente
    UuidRef:
      type: object
      required:
        - publicId
      properties:
        publicId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []