# backend/src/main/java/br/com/salao/web/dto/ServiceResponse.java
package br.com.salao.web.dto;

import java.math.BigDecimal;
import java.util.UUID;

public record ServiceResponse(
        UUID publicId,
        String name,
        String description,
        Integer durationMinutes,
        BigDecimal price,
        boolean active
) {
}